Enhancements and documentation fixes.
