--- scripts/cpu/processors.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-threads=$(cat /proc/cpuinfo | grep processor | wc -l)
-echo "KVAL thread_count \"${threads}\""
-
